backend/requirements.txt
fastapi
requests
pillow
numpy
uvicorn